Expanding the Test Cases for the DHT to include coverage of all the message types and the Message Factory.  Removing unused code and cleaning up the coding style of the classes so they are consistent.

+++ dht/src/main/java/io/topiacoin/dht/messages/NodeLookupResponse.java
@@ -17,7 +17,7 @@ public class NodeLookupResponse implements Message {
 
     private final Log _log = LogFactory.getLog(this.getClass());
 
-    public static final int TYPE = (byte)0x83;
+    public static final byte TYPE = (byte)0x83;
 
     private List<Node> nodes;
